Remove MMPL-licensed mods
http://www.mod-buildcraft.com/MMPL-1.0.txt Minecraft Mod Public License
contains the following provision:
"user must own a legally acquired copy of Minecraft"
which seems to (disclaimer: IANAL) conflict with
http://opensource.org/osd Open Source Initiative OSD
"8. License Must Not Be Specific to a Product
The rights attached to the program must not depend on the program's
being part of a particular software distribution. If the program is
extracted from that distribution and used or distributed within the
terms of the program's license, all parties to whom the program is
redistributed should have the same rights as those that are granted in
conjunction with the original software distribution"
and this license is not listed on http://opensource.org/licenses/
(i.e., hasn't gone through the review process). For this reason,
removing the MMPL-licensed mods for now.